feat: #100 카테고리 버전별 카테고리 DTO 조회 쿼리 구현

+++ src/main/java/com/waguwagu/weat/domain/category/model/entity/CategoryTag.java
@@ -29,7 +29,7 @@ public class CategoryTag {
 
     @Column(name = "category_tag_order")
     @Comment("카테고리 태그순서")
-    private Long categoryTagOrder;
+    private Integer categoryTagOrder;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "category_id", foreignKey = @ForeignKey(name = "category_tag_category_fk"))
